Improved MIASequencer code organization
