+++ bin/apps/MIASequencer/CMakeLists.txt
@@ -1,11 +1,13 @@
 # Create the MIASequencer executable.
 set(MIASequencer_SRC 
-    MIASequencer.cpp 
-    MIASequencer_main.cpp 
-    KeyListenerTask.cpp )
+    MIASequencer.cpp
+    MIASequencer_main.cpp
+    KeyListenerTask.cpp
+    Sequences.cpp )
 set(MIASequencer_INC 
-    MIASequencer.hpp 
-    KeyListenerTask.hpp )
+    MIASequencer.hpp
+    KeyListenerTask.hpp
+    Sequences.hpp )
 add_executable(MIASequencer ${MIASequencer_SRC} ${MIASequencer_INC} )
 target_link_libraries(MIASequencer PRIVATE Framework_CORE System_UTIL Types_UTIL Math_UTIL )
