+++ docs/MIAManual.tex
@@ -2,7 +2,7 @@
 
 % Define the version number here
 % This is automatically incremented when using the commitAll.sh script.
-\newcommand{\version}{Version 2.067}
+\newcommand{\version}{Version 2.068}
 
 \include{./TeX_files/packages}
 \include{./TeX_files/settings}
